+++ packages/ui/src/components/sections/openchamber/DesktopNetworkSettings.tsx
@@ -14,40 +14,24 @@ import {
   setDesktopKeepAwake,
   setDesktopLaunchAtLogin,
   setDesktopMinimizeToTray,
-  usesFramelessElectronChrome,
-  type DesktopWindowControlsPosition,
 } from '@/lib/desktop';
 import { useI18n } from '@/lib/i18n';
-import { updateDesktopSettings } from '@/lib/persistence';
 import { runtimeFetch } from '@/lib/runtime-fetch';
 import { getRuntimeApiBaseUrl } from '@/lib/runtime-switch';
-import { useUIStore } from '@/stores/useUIStore';
 import {
   SettingsSection,
   SettingsCheckboxRow,
-  SettingsChipGroup,
-  SettingsFieldRow,
   SETTINGS_OPTION_STACK_CLASS,
   SettingsStackedField,
   SETTINGS_ICON_BUTTON_CLASS,
 } from '@/components/sections/shared/SettingsSection';
 
-const WINDOW_CONTROLS_POSITION_OPTIONS: Array<{ id: DesktopWindowControlsPosition; labelKey: string }> = [
-  { id: 'auto', labelKey: 'settings.openchamber.desktopNetwork.option.windowControlsAuto' },
-  { id: 'left', labelKey: 'settings.openchamber.desktopNetwork.option.windowControlsLeft' },
-  { id: 'right', labelKey: 'settings.openchamber.desktopNetwork.option.windowControlsRight' },
-];
-
 export const DesktopNetworkSettings: React.FC = () => {
   const { t } = useI18n();
-  const tUnsafe = React.useCallback((key: string) => t(key as Parameters<typeof t>[0]), [t]);
   const isLocalDesktop = isDesktopShell() && isDesktopLocalOriginActive();
   const isMacDesktop = isLocalDesktop
     && typeof window !== 'undefined'
     && window.__OPENCHAMBER_PLATFORM__ === 'darwin';
-  const showWindowControlsPosition = usesFramelessElectronChrome();
-  const desktopWindowControlsPosition = useUIStore((state) => state.desktopWindowControlsPosition);
-  const setDesktopWindowControlsPosition = useUIStore((state) => state.setDesktopWindowControlsPosition);
   const [savedValue, setSavedValue] = React.useState(false);
   const [draftValue, setDraftValue] = React.useState(false);
   const [savedPassword, setSavedPassword] = React.useState('');
@@ -242,11 +226,6 @@ export const DesktopNetworkSettings: React.FC = () => {
     }
   }, []);
 
-  const handleWindowControlsPositionChange = React.useCallback((value: DesktopWindowControlsPosition) => {
-    setDesktopWindowControlsPosition(value);
-    void updateDesktopSettings({ desktopWindowControlsPosition: value });
-  }, [setDesktopWindowControlsPosition]);
-
   const handleLaunchAtLoginToggle = React.useCallback(async () => {
     if (!launchAtLoginSupported || isSavingLaunchAtLogin) {
       return;
@@ -362,181 +341,156 @@ export const DesktopNetworkSettings: React.FC = () => {
     }
   }, [draftMacMenuBarEnabled, draftPassword, draftValue, isDirty, t]);
 
-  if (!isLocalDesktop && !showWindowControlsPosition) {
+  if (!isLocalDesktop) {
     return null;
   }
 
   return (
-    <>
-      {showWindowControlsPosition ? (
-        <SettingsSection title={t('settings.openchamber.desktopNetwork.field.windowControlsPosition')}>
-          <SettingsFieldRow
-            settingsItem="sessions.desktop-window-controls-position"
-            label={t('settings.openchamber.desktopNetwork.field.windowControlsPositionDescription')}
-            alignEnd={false}
-            controlClassName="flex-col items-stretch"
-          >
-            <SettingsChipGroup
-              value={desktopWindowControlsPosition}
-              options={WINDOW_CONTROLS_POSITION_OPTIONS.map((option) => ({
-                value: option.id,
-                label: tUnsafe(option.labelKey),
-              }))}
-              onChange={handleWindowControlsPositionChange}
-              aria-label={t('settings.openchamber.desktopNetwork.field.windowControlsPositionAria')}
-            />
-          </SettingsFieldRow>
-        </SettingsSection>
-      ) : null}
-
-      {isLocalDesktop ? (
-        <SettingsSection title={t('settings.openchamber.desktopNetwork.title')}>
-          <div className="space-y-3">
-            {(launchAtLoginSupported || isMacDesktop || minimizeToTraySupported || keepAwakeSupported) ? (
-              <div className={SETTINGS_OPTION_STACK_CLASS}>
-                {launchAtLoginSupported ? (
-                  <SettingsCheckboxRow
-                    settingsItem="sessions.desktop-launch-at-login"
-                    checked={launchAtLoginEnabled}
-                    onChange={(checked) => {
-                      if (checked === launchAtLoginEnabled) return;
-                      void handleLaunchAtLoginToggle();
-                    }}
-                    disabled={isSavingLaunchAtLogin}
-                    label={t('settings.openchamber.desktopNetwork.field.launchAtLogin')}
-                    info={t('settings.openchamber.desktopNetwork.field.launchAtLoginDescription')}
-                    ariaLabel={t('settings.openchamber.desktopNetwork.field.launchAtLoginAria')}
-                  />
-                ) : null}
-
-                {isMacDesktop ? (
-                  <SettingsCheckboxRow
-                    settingsItem="sessions.desktop-mac-menu-bar"
-                    checked={draftMacMenuBarEnabled}
-                    onChange={setDraftMacMenuBarEnabled}
-                    disabled={isLoading || isSaving}
-                    label={t('settings.openchamber.desktopNetwork.field.macMenuBar')}
-                    info={t('settings.openchamber.desktopNetwork.field.macMenuBarDescription')}
-                    ariaLabel={t('settings.openchamber.desktopNetwork.field.macMenuBarAria')}
-                  />
-                ) : null}
-
-                {minimizeToTraySupported ? (
-                  <SettingsCheckboxRow
-                    settingsItem="sessions.desktop-minimize-to-tray"
-                    checked={minimizeToTrayEnabled}
-                    onChange={(checked) => {
-                      if (checked === minimizeToTrayEnabled) return;
-                      void handleMinimizeToTrayToggle();
-                    }}
-                    disabled={isSavingMinimizeToTray}
-                    label={t('settings.openchamber.desktopNetwork.field.minimizeToTray')}
-                    info={t('settings.openchamber.desktopNetwork.field.minimizeToTrayDescription')}
-                    ariaLabel={t('settings.openchamber.desktopNetwork.field.minimizeToTrayAria')}
-                  />
-                ) : null}
-
-                {keepAwakeSupported ? (
-                  <SettingsCheckboxRow
-                    settingsItem="sessions.desktop-keep-awake"
-                    checked={keepAwakeEnabled}
-                    onChange={(checked) => {
-                      if (checked === keepAwakeEnabled) return;
-                      void handleKeepAwakeToggle();
-                    }}
-                    disabled={isSavingKeepAwake}
-                    label={t('settings.openchamber.desktopNetwork.field.keepAwake')}
-                    info={t('settings.openchamber.desktopNetwork.field.keepAwakeDescription')}
-                    ariaLabel={t('settings.openchamber.desktopNetwork.field.keepAwakeAria')}
-                  />
-                ) : null}
-              </div>
-            ) : null}
-
-            <SettingsStackedField
-              settingsItem="sessions.desktop-ui-password"
-              label={(
-                <label htmlFor="desktop-ui-password">
-                  {t('settings.openchamber.desktopPassword.field.password')}
-                </label>
-              )}
-              info={t('settings.openchamber.desktopPassword.field.passwordDescription')}
-            >
-              <Input
-                id="desktop-ui-password"
-                type={showPassword ? 'text' : 'password'}
-                className="h-8 min-w-0 flex-1"
-                value={draftPassword}
-                onChange={(event) => handlePasswordChange(event.target.value)}
-                placeholder={t('settings.openchamber.desktopPassword.field.passwordPlaceholder')}
-                disabled={isLoading || isSaving}
-                required={draftValue}
-                aria-invalid={lanRequiresPassword}
-              />
-              <Button
-                type="button"
-                variant="ghost"
-                size="xs"
-                onClick={() => setShowPassword((current: boolean) => !current)}
-                className={SETTINGS_ICON_BUTTON_CLASS}
-                aria-label={t(showPassword ? 'settings.openchamber.desktopPassword.actions.hidePassword' : 'settings.openchamber.desktopPassword.actions.showPassword')}
-                aria-pressed={showPassword}
-              >
-                <Icon name={showPassword ? 'eye-off' : 'eye'} className="h-4 w-4" />
-              </Button>
-            </SettingsStackedField>
-
-            <div className={SETTINGS_OPTION_STACK_CLASS}>
+    <SettingsSection title={t('settings.openchamber.desktopNetwork.title')}>
+      <div className="space-y-3">
+        {(launchAtLoginSupported || isMacDesktop || minimizeToTraySupported || keepAwakeSupported) ? (
+          <div className={SETTINGS_OPTION_STACK_CLASS}>
+            {launchAtLoginSupported ? (
               <SettingsCheckboxRow
-                settingsItem="sessions.desktop-lan-access"
-                checked={draftValue}
-                onChange={setDraftValue}
-                disabled={isLoading || isSaving}
-                label={t('settings.openchamber.desktopNetwork.field.allowLanAccess')}
-                info={t('settings.openchamber.desktopNetwork.field.allowLanAccessDescription')}
-                description={(
-                  <>
-                    <span className="block text-[var(--status-warning)]/85">
-                      {t('settings.openchamber.desktopNetwork.field.warning')}
-                    </span>
-                    {lanRequiresPassword || lanBlockedByMissingPassword ? (
-                      <span className="block text-[var(--status-warning)]/85">
-                        {t('settings.openchamber.desktopNetwork.field.passwordRequiredWarning')}
-                      </span>
-                    ) : null}
-                  </>
-                )}
-                ariaLabel={t('settings.openchamber.desktopNetwork.field.allowLanAccessAria')}
+                settingsItem="sessions.desktop-launch-at-login"
+                checked={launchAtLoginEnabled}
+                onChange={(checked) => {
+                  if (checked === launchAtLoginEnabled) return;
+                  void handleLaunchAtLoginToggle();
+                }}
+                disabled={isSavingLaunchAtLogin}
+                label={t('settings.openchamber.desktopNetwork.field.launchAtLogin')}
+                info={t('settings.openchamber.desktopNetwork.field.launchAtLoginDescription')}
+                ariaLabel={t('settings.openchamber.desktopNetwork.field.launchAtLoginAria')}
               />
-            </div>
-
-            {error ? (
-              <div className="typography-micro text-[var(--status-error)]">{error}</div>
             ) : null}
 
-            {lanUrl ? (
-              <div className="typography-micro text-muted-foreground/80">
-                {isDirty && !savedValue
-                  ? t('settings.openchamber.desktopNetwork.hint.openAfterRestart')
-                  : t('settings.openchamber.desktopNetwork.hint.openNow')}
-                <span className="font-mono text-foreground">{lanUrl}</span>
-              </div>
+            {isMacDesktop ? (
+              <SettingsCheckboxRow
+                settingsItem="sessions.desktop-mac-menu-bar"
+                checked={draftMacMenuBarEnabled}
+                onChange={setDraftMacMenuBarEnabled}
+                disabled={isLoading || isSaving}
+                label={t('settings.openchamber.desktopNetwork.field.macMenuBar')}
+                info={t('settings.openchamber.desktopNetwork.field.macMenuBarDescription')}
+                ariaLabel={t('settings.openchamber.desktopNetwork.field.macMenuBarAria')}
+              />
             ) : null}
 
-            <div className="flex justify-start py-1.5">
-              <Button
-                type="button"
-                size="xs"
-                onClick={handleSaveAndRestart}
-                disabled={saveDisabled}
-                className="shrink-0 !font-normal"
-              >
-                {isSaving ? t('settings.common.actions.saving') : t('settings.openchamber.desktopNetwork.actions.saveAndRestart')}
-              </Button>
-            </div>
+            {minimizeToTraySupported ? (
+              <SettingsCheckboxRow
+                settingsItem="sessions.desktop-minimize-to-tray"
+                checked={minimizeToTrayEnabled}
+                onChange={(checked) => {
+                  if (checked === minimizeToTrayEnabled) return;
+                  void handleMinimizeToTrayToggle();
+                }}
+                disabled={isSavingMinimizeToTray}
+                label={t('settings.openchamber.desktopNetwork.field.minimizeToTray')}
+                info={t('settings.openchamber.desktopNetwork.field.minimizeToTrayDescription')}
+                ariaLabel={t('settings.openchamber.desktopNetwork.field.minimizeToTrayAria')}
+              />
+            ) : null}
+
+            {keepAwakeSupported ? (
+              <SettingsCheckboxRow
+                settingsItem="sessions.desktop-keep-awake"
+                checked={keepAwakeEnabled}
+                onChange={(checked) => {
+                  if (checked === keepAwakeEnabled) return;
+                  void handleKeepAwakeToggle();
+                }}
+                disabled={isSavingKeepAwake}
+                label={t('settings.openchamber.desktopNetwork.field.keepAwake')}
+                info={t('settings.openchamber.desktopNetwork.field.keepAwakeDescription')}
+                ariaLabel={t('settings.openchamber.desktopNetwork.field.keepAwakeAria')}
+              />
+            ) : null}
           </div>
-        </SettingsSection>
-      ) : null}
-    </>
+        ) : null}
+
+        <SettingsStackedField
+          settingsItem="sessions.desktop-ui-password"
+          label={(
+            <label htmlFor="desktop-ui-password">
+              {t('settings.openchamber.desktopPassword.field.password')}
+            </label>
+          )}
+          info={t('settings.openchamber.desktopPassword.field.passwordDescription')}
+        >
+          <Input
+            id="desktop-ui-password"
+            type={showPassword ? 'text' : 'password'}
+            className="h-8 min-w-0 flex-1"
+            value={draftPassword}
+            onChange={(event) => handlePasswordChange(event.target.value)}
+            placeholder={t('settings.openchamber.desktopPassword.field.passwordPlaceholder')}
+            disabled={isLoading || isSaving}
+            required={draftValue}
+            aria-invalid={lanRequiresPassword}
+          />
+          <Button
+            type="button"
+            variant="ghost"
+            size="xs"
+            onClick={() => setShowPassword((current: boolean) => !current)}
+            className={SETTINGS_ICON_BUTTON_CLASS}
+            aria-label={t(showPassword ? 'settings.openchamber.desktopPassword.actions.hidePassword' : 'settings.openchamber.desktopPassword.actions.showPassword')}
+            aria-pressed={showPassword}
+          >
+            <Icon name={showPassword ? 'eye-off' : 'eye'} className="h-4 w-4" />
+          </Button>
+        </SettingsStackedField>
+
+        <div className={SETTINGS_OPTION_STACK_CLASS}>
+          <SettingsCheckboxRow
+            settingsItem="sessions.desktop-lan-access"
+            checked={draftValue}
+            onChange={setDraftValue}
+            disabled={isLoading || isSaving}
+            label={t('settings.openchamber.desktopNetwork.field.allowLanAccess')}
+            info={t('settings.openchamber.desktopNetwork.field.allowLanAccessDescription')}
+            description={(
+              <>
+                <span className="block text-[var(--status-warning)]/85">
+                  {t('settings.openchamber.desktopNetwork.field.warning')}
+                </span>
+                {lanRequiresPassword || lanBlockedByMissingPassword ? (
+                  <span className="block text-[var(--status-warning)]/85">
+                    {t('settings.openchamber.desktopNetwork.field.passwordRequiredWarning')}
+                  </span>
+                ) : null}
+              </>
+            )}
+            ariaLabel={t('settings.openchamber.desktopNetwork.field.allowLanAccessAria')}
+          />
+        </div>
+
+        {error ? (
+          <div className="typography-micro text-[var(--status-error)]">{error}</div>
+        ) : null}
+
+        {lanUrl ? (
+          <div className="typography-micro text-muted-foreground/80">
+            {isDirty && !savedValue
+              ? t('settings.openchamber.desktopNetwork.hint.openAfterRestart')
+              : t('settings.openchamber.desktopNetwork.hint.openNow')}
+            <span className="font-mono text-foreground">{lanUrl}</span>
+          </div>
+        ) : null}
+
+        <div className="flex justify-start py-1.5">
+          <Button
+            type="button"
+            size="xs"
+            onClick={handleSaveAndRestart}
+            disabled={saveDisabled}
+            className="shrink-0 !font-normal"
+          >
+            {isSaving ? t('settings.common.actions.saving') : t('settings.openchamber.desktopNetwork.actions.saveAndRestart')}
+          </Button>
+        </div>
+      </div>
+    </SettingsSection>
   );
 };
